+++ crates/sncast/src/lib.rs
@@ -127,30 +127,6 @@ pub struct AccountData {
     pub account_type: Option<AccountType>,
 }
 
-#[derive(Debug, PartialEq, Eq, Copy, Clone, Serialize)]
-pub enum NumbersFormat {
-    Default,
-    Decimal,
-    Hex,
-}
-
-impl NumbersFormat {
-    #[must_use]
-    pub fn from_flags(hex_format: bool, dec_format: bool) -> Self {
-        assert!(
-            !(hex_format && dec_format),
-            "Exclusivity should be validated by clap"
-        );
-        if hex_format {
-            NumbersFormat::Hex
-        } else if dec_format {
-            NumbersFormat::Decimal
-        } else {
-            NumbersFormat::Default
-        }
-    }
-}
-
 #[derive(Clone, Copy)]
 pub struct WaitForTx {
     pub wait: bool,