+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - bug where syscall execution resources from nested calls were being calculated twice
 
+### Cast
+
+#### Removed
+
+- `--int-format` and `--hex-format`, all values are displayed with default format
+
 ## [0.45.0] - 2025-06-16
 
 ### Forge